+++ src/main/java/io/github/talelin/merak/modules/message/MessageWebSocketHandler.java
@@ -1,4 +1,4 @@
-package io.github.talelin.merak.extensions.message;
+package io.github.talelin.merak.modules.message;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.socket.*;
